# webserver/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "0.0.2",
        "title": "Linto admin API documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@linto.ai"
        }
    },
    "host": "linto.ai/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/api/androidusers": {
            "get": {
                "tags": [
                    "android_users"
                ],
                "summary": "Get all andoird users",
                "operationId": "getAllAndroidUsers",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "applications": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "android_users"
                ],
                "summary": "Add a new andoird users",
                "operationId": "addAndroidUsers",
                "parameters": [{
                    "in": "body",
                    "name": "payload",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string",
                                "format": "email"
                            },
                            "pswd": {
                                "type": "string"
                            },
                            "applications": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }

                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
        "/api/androidusers/applications": {
            "patch": {
                "tags": [
                    "android_users"
                ],
                "summary": "Remove an application for all android users",
                "operationId": "removeApplicationFromAndroidUsers",
                "parameters": [{
                    "name": "payload",
                    "in": "body",
                    "required": "true",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }

                }
            }
        },
        "/api/androidusers/{userId}/applications": {
            "put": {
                "tags": [
                    "android_users"
                ],
                "summary": "Register an android user to an application",
                "operationId": "AddApplicationToAndroidUser",
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "Android user id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "payload",
                    "in": "body",
                    "description": "Array of application to add to android user",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "applications": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "enum": ["success", "error"],
                                    "required": true
                                },
                                "msg": {
                                    "type": "string",
                                    "required": true
                                },
                                "error": {
                                    "type": "object",
                                    "required": false
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/androidusers/{userId}/applications/{applicationId}/remove": {
            "patch": {
                "tags": [
                    "android_users"
                ],
                "summary": "Dissociate an android user from an android application",
                "operationId": "RemoveApplicationFromAndroidUser",
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "Android user id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "applicationId",
                    "in": "path",
                    "description": "Application workflow id to remove",
                    "required": true,
                    "type": "string"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
        "/api/androidusers/{userId}": {
            "get": {
                "tags": [
                    "android_users"
                ],
                "summary": "Get an android user by its id",
                "operationId": "GetAndroidUserById",
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "Android user id",
                    "required": true,
                    "type": "string"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {

                            "type": "object",
                            "properties": {
                                "_id": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email"
                                },
                                "applications": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "android_users"
                ],
                "summary": "Delete an android user by its id",
                "operationId": "deleteAndroidUser",
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "Android user id",
                    "required": true,
                    "type": "string"
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
        "/api/workflows/application/{id}": {
            "get": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Get an application workflow by its id",
                "operationId": "GetApplicationWorkflowById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "required": true,
                    "name": "id",
                    "description": "Application workflow id",
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/workflows_application"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Delete an application workflow by its id",
                "operationId": "DeleteApplicationWorkflow",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "path",
                        "required": true,
                        "name": "workflowId",
                        "description": "Application workflow id",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "workflowName",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/workflows_application"
                        }
                    }
                }
            }
        },
        "/api/workflows/application/{id}/androidusers": {
            "get": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Get a user list associated to application {id}",
                "operationId": "GetAndroidUserByApplication",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "required": true,
                    "name": "id",
                    "description": "Application workflow id",
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "applications": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "workflows_application": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "flowId": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "updated_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "flow": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "label": {
                            "type": "string"
                        },
                        "nodes": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        },
                        "configs": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
        "server_response": {

            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "enum": ["success", "error"],
                    "required": true
                },
                "msg": {
                    "type": "string",
                    "required": true
                },
                "error": {
                    "type": "object",
                    "required": false
                }
            }
        }
    }
}